make the survey list item a component
